+++ tests/test_async_s3_bucket.py
@@ -60,7 +60,15 @@ class AsyncS3BucketTest(unittest.IsolatedAsyncioTestCase):
             assert bucket.service_resource.meta.client.meta.endpoint_url == endpoint_url
 
     async def test_file_transfer(self) -> None:
-        await self.bucket.upload(fileobj=self.fileobj, key=self.key)
+        truth_lfoa_url = "/".join(
+            [
+                self.bucket.service_resource.meta.client.meta.endpoint_url,
+                self.bucket_name,
+                self.key,
+            ]
+        )
+        lfoa_url = await self.bucket.upload(fileobj=self.fileobj, key=self.key)
+        assert lfoa_url == truth_lfoa_url
         roundtrip_fileobj = await self.bucket.download(key=self.key)
         roundtrip_data = roundtrip_fileobj.read()
         assert self.file_data == roundtrip_data
@@ -219,6 +227,18 @@ class AsyncS3BucketClassmethodTest(unittest.IsolatedAsyncioTestCase):
             "Foo:5/testFiberSpecBlue/2020/04/02/"
             "Foo:5_testFiberSpecBlue_2020-04-02T12:00:00.000suffixtext"
         )
+        assert key == expected_key
+
+        # Repeat the test with a specified value for "stem"
+        key = salobj.AsyncS3Bucket.make_key(
+            salname=salname,
+            salindexname=5,
+            generator=generator,
+            date=date,
+            stem="FF_O_20200402_000001",
+            suffix=".fits",
+        )
+        expected_key = "Foo:5/testFiberSpecBlue/2020/04/02/FF_O_20200402_000001.fits"
         assert key == expected_key
 
     def test_env_var_secrets(self) -> None: